requirements.txt
fastapi
uvicorn
google-api-python-client
google-auth
google-auth-oauthlib
google-cloud-firestore
google-cloud-secret-manager
google-cloud-aiplatform
python-dotenv